Initial request and response parsing working
Still need to account for chunked responses. Also need to understand
why `setResponse` is only working when the `request` is instantiated in
`setRequestSocket`, necessitating the ugly need for
`this._setResponseResolver`. It seems like this should work when
instantiated in `setResponse` where it can call the promise's `resolve()`
directly, but it does not.